% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity_2.R
\name{ethnic_code_long}
\alias{ethnic_code_long}
\title{Code ethnicity as per Statistics NZ guidelines: Long}
\usage{
ethnic_code_long(
  data,
  indicator_cols,
  indicator_codes = census_2013_question_codes,
  text_cols,
  text_delim = ",",
  text_code_level = 4,
  level_out = 3
)
}
\arguments{
\item{data}{A data frame containing ethnicity questions to code.}

\item{indicator_cols}{Columns containing ethnicity indicator columns}

\item{indicator_codes}{A vector containing the StatsNZ codes for the ethnicity indicator columns. The length of codes should match the number of columns listed, and be in the same order. Default is as per the standard New Zealand census ethnicity question (used in 2001, 2006, 2013, and 2018). I.e. New Zealand European = 111, Māori = 211, ... .}

\item{text_cols}{<tidy-select> Columns containing ethnicity text or numeric columns}

\item{text_delim}{Provided deliminator if multiple ethnicities are recorded in a single column with deliminator separating them.}

\item{text_code_level}{Provide the StatsNZ classification level (1, 2, 3, or 4) that ethnicity has been entered as.}

\item{level_out}{Provide the StatsNZ classification level (1, 2, 3, or 4) that ethnicity is to be outputted as.}
}
\description{
Code ethnicity as per Statistics NZ guidelines: Long
}
