% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity_2.R
\name{ethnic_code_text}
\alias{ethnic_code_text}
\title{Provides StatsNZ codes for ethnicity recorded in text or numeric columns}
\usage{
ethnic_code_text(
  data,
  cols,
  delim = ",",
  code_level = 4,
  type = c("text", "code"),
  check = FALSE
)
}
\arguments{
\item{data}{A data frame containing ethnicity questions to code.}

\item{cols}{<tidy-select> Columns containing ethnicity text or numeric columns}

\item{delim}{Provided deliminator if multiple ethnicities are recorded in a single column with deliminator separating them.}

\item{code_level}{Provide the StatsNZ classification level (1,2,3 or 4) that ethnicity has been entered as.}

\item{type}{Indicate whether ethnicity has been recorded as text (e.g. Taiwanese) or StatsNZ numeric code (e.g. 42116)}

\item{check}{Logical to determine whether to output coded dataset (check = FALSE) or only return uncoded ethnicities (check = TRUE).}
}
\description{
Provides StatsNZ codes for ethnicity recorded in text or numeric columns
}
