% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity_2.R
\name{ethnic_code_wide}
\alias{ethnic_code_wide}
\title{Code ethnicity as per Statistics NZ guidelines: Wide}
\usage{
ethnic_code_wide(data, level_out = 1, col_names = ethnicity_level1_codes)
}
\arguments{
\item{data}{data frame containing ethnicity questions to code.}

\item{level_out}{Provide the StatsNZ classification level (1, 2, 3, or 4) that ethnicity is to be outputted as.}

\item{col_names}{Names of output columns.}
}
\description{
Code ethnicity as per Statistics NZ guidelines: Wide
}
