% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_internal_data.R
\name{find_data}
\alias{find_data}
\title{Check for internal data}
\usage{
find_data()
}
\value{
A data frame with the list of files that can be accessed through the NHANES website.  Should not generally be used.  Present for debugging purposes and transparency.
}
\description{
Provides access to the internal data listing all NHANES files
}
