% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity.R
\name{tidy_ethnicity}
\alias{tidy_ethnicity}
\title{Code ethnicity as per Statistics NZ guidelines.}
\usage{
tidy_ethnicity(
  data,
  cols = 1:10,
  sep = ",",
  base_levels = NULL,
  eth_levels = NULL,
  eth_prior = NULL,
  prior_order = c(2:6, 1, 7),
  add_cols = FALSE
)
}
\arguments{
\item{data}{Data collected as per the standard New Zealand census ethnicity
question (used in 2001, 2006, 2013, and 2018).}

\item{cols}{Vector of numeric indices or column names of length 10,
indicating positions or names of ethnicity variables. Columns should be in
the following order (names aren't important); New Zealand European, Maori,
Samoan, Cook Island Maori, Tongan, Niuean, Chinese, Indian, Other, Other
specified. First nine columns should be logical or coercable as such (coded
as 0 = No or 1 = Yes). Last column should be a text field with multiple
ethnicities separated by a character string.}

\item{sep}{A character string used to separate different ethnicities listed in
the 'Other' field. Comma by default.}

\item{base_levels}{Names of input variables as recorded in the reference
dataset although not necessary as recorded in the data. Defaults to New
Zealand European, Maori, Samoan, Cook Island Maori, Tongan, Niuean,
Chinese, Indian, Other}

\item{eth_levels}{Names of output columns. Defaults to European, Maori,
Pacific, Asian, MELAA, Other, and Unknown.}

\item{eth_prior}{A character string indicating the name to give a column
containing prioritised ethnicity. Default is not to produce column.}

\item{prior_order}{Numeric vector giving prioritisation order of columns
listed in eth_levels. Defaults to c(2:6,1,7), ie. Maori, Pacific, Asian,
MELAA, Other, European, and Unknown.}

\item{add_cols}{Either a logical vector indicating whether to append the
output to data collected, or a character vector containing the names of one
or more variables to include in the output. Default is to return just the
calculated ethnicity indicator variables.}
}
\value{
A data.frame with ethnicity formatted according to Statistics NZ
  levels.
}
\description{
Code ethnicity as per Statistics NZ guidelines.
}
