% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity.R
\name{tidy_ethnicity_codes}
\alias{tidy_ethnicity_codes}
\title{Code ethnicity recorded as Statistics NZ codes}
\usage{
tidy_ethnicity_codes(
  data,
  vars_prefix,
  vars_binary,
  vars_other,
  level_out = 1,
  col_names = paste0("eth_", c("euro", "maori", "pacific", "asian", "melaa", "other",
    "unknown")),
  eth_prior = NULL,
  eth_levels = NULL,
  prior_order = c(2:6, 1, 7),
  add_cols = TRUE
)
}
\arguments{
\item{data}{Data collected as per the standard New Zealand census ethnicity
question (used in 2001, 2006, 2013, and 2018). 
Ethnicity columns should be named with a common prefix followed by valid StatsNZ ethnicity codes
(e.g. eth___11, eth___21, ...).}

\item{vars_prefix}{Common prefix used to denote ethnicity variable names.}

\item{vars_binary}{Tidy-select argument to select ethnicity columns. 
Ethnicity columns should be named with a common prefix followed by valid StatsNZ ethnicity codes
  (e.g. eth___11, eth___21, ...). 
 Columns should be logical or coercable as such (coded as 0 = No or 1 = Yes).}

\item{vars_other}{Tidy-select argument to select columns containing 'other' ethnicities recording StatsNZ codes.}

\item{level_out}{Integer indicating StatsNZ total response output level (1, 2, 3 or 4).}

\item{col_names}{Names to give output columns.}

\item{eth_prior}{A character string indicating the name to give a column
containing prioritised ethnicity. Default is not to produce column.}

\item{eth_levels}{Names of output columns. Defaults to European, Maori,
Pacific, Asian, MELAA, Other, and Unknown.}

\item{prior_order}{Numeric vector giving prioritisation order of columns
listed in eth_levels. Defaults to c(2:6,1,7), ie. Maori, Pacific, Asian,
MELAA, Other, European, and Unknown.}

\item{add_cols}{Either a logical vector indicating whether to append the
output to data collected, or a character vector containing the names of one
or more variables to include in the output. Default is to return just the
calculated ethnicity indicator variables.}
}
\value{
A data.frame with ethnicity formatted according to Statistics NZ
  levels.
}
\description{
Code ethnicity recorded as Statistics NZ codes
}
