% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZEthnicity.R
\name{tidy_ethnicity}
\alias{tidy_ethnicity}
\title{Title}
\usage{
tidy_ethnicity(
  data,
  cols = 1:10,
  sep = ",",
  id_col = NULL,
  base_levels = c("New Zealand European", "Maori", "Samoan", "Cook Island Maori",
    "Tongan", "Niuean", "Chinese", "Indian", "Other"),
  eth_levels = c("European", "Maori", "Pacific", "Asian", "MELAA", "Other", "Residual")
)
}
\arguments{
\item{data}{Data collected as per the standard New Zealand census ethnicity
question (used in 2001, 2006, 2013, and 2018).}

\item{cols}{Vector of numeric or character indices indicating positions or
names of ethnicity variables. Columns should be in the following order
(names aren't important); New Zealand European, Maori, Samoan, Cook Island
Maori, Tongan, Niuean, Chinese, Indian, Other, Other specified, (Refused)
Columns should be logical or coercable as such (coded as 0 = No or 1 =
Yes). Other specified should be a text field with multiple ethnicities
separated by a comma.}

\item{sep}{A character string used to seperate different ethnicities list in the 'Other' field.}

\item{base_levels}{Names of input variables, as recorded in the reference dataset.}

\item{eth_levels}{Names of output columns.}
}
\value{
A data.frame with ethnicity formatted according to Statistics NZ levels.
}
\description{
Title
}
